### 2020-02-01 #2058 @Kevin-Mattheus-Moerman
@telatin thanks for your submission. I recommend that you reformulate your readme to be a bit more suitable to somebody from outside of this domain (like myself). It contains 2 undefined acronyms for instance. Perhaps create a lay-persons summary at the top so it is clear what this software is for to new comers.
Furthermore if this work is reviewed at a later stage our reviewers will demand clear contributing guidelines. I have not found these from quickly looking through the readme. Perhaps you can consider adding clear contributing guidelines at this point while we determine the suitability of this software for JOSS. Thanks.
