### 2020-02-02 18cedbc
Create CONTRIBUTING.md
