+++ CONTRIBUTING.md
@@ -1,0 +1,6 @@
+If you wish to fix a bug or add new features to the software we welcome Pull Requests. 
+
+We use [GitHub Flow style development](https://guides.github.com/introduction/flow/). 
+
+Please fork the repo, make the change, then submit a Pull Request against out master branch, with details about what the change is and what it fixes/adds. 
+We will then review your changes and merge them, or provide feedback on enhancements.
