### 2020-02-03 #2058 @will-rowe
Good morning. I've been looking at this submission this morning and I'm afraid we may class this as a minor-utility. Whilst I think this is a useful tool and one which I would use to save myself bedtools + awk scripting, I don't think there is enough functionality here to pass our submission requirements.

I'd appreciate it if either @lpantano, @majensen or @csoneson could take a quick look and confirm this.

Best,

Will
