### 2020-02-04 #2058 @lpantano
It seems a nice tool, if it is actually super fast compare to others like bedtools or mosdepth, it could be a good improvement. Otherwise, I could agree is minor.
