### 2020-02-05 #2058 @telatin
Thanks @will-rowe and @lpantano  for your first feedback!

We were aware of this possible limitation, yet we were also thinking that "covtobed" fits into a broad set of common tasks, and different bioinformatics applications this specificity was somehow desirable. So we focused on supporting streaming and speed to have a reliable brick for multiple pipelines.
