### 2020-02-05 #2058 @will-rowe
Hi @telatin. Thanks for your reply. I've been playing with covtobed again this morning - it is a good utility and is very nicely packaged and documented.

Of course you and @lpantano are right in that the speed makes this appealing. I was testing against bedtools (`bedtools genomecov -ibam a.bam -bg`), which gives the same output to `covtobed a.bam`. covtobed was roughly twice as fast (note: there were some output inconsistencies but that seems to be due to just using defaults for both programs) and allows streaming (which bedtools doesn't for genomecov at the moment).

Maybe if you could add some benchmarking to your paper and that would satisfy our concerns that this is a minor utility?

I have an alternative (and somewhat cheeky suggestion). Adding some functionality to generate a coverage plot from the bed file would make this an indispensable tool to many bioinformaticians. Something along the lines of the image in your README would be fantastic. This is of course only a suggestion and I'd be happy to put this paper into review with only the benchmarking data added
