### 2020-02-05 #2058 @telatin
@will-rowe I really appreciate the time you dedicated to this tool!

To be fair, both suggestions are valuable, and I'll be glad to try addressing one or both.
This will probably take some time, is it better to close this issue and resubmit mentioning the issue later or is it possible to keep the issue open in the meantime?
