+++ benchmark/benchmark_example2.md
@@ -1,0 +1,4 @@
+| Command | Mean [s] | Min [s] | Max [s] | Relative |
+|:---|---:|---:|---:|---:|
+| `covtobed /local/giovanni/covtobed/benchmark/../TE/example2.bam > /dev/null` | 1.863 ± 0.015 | 1.854 | 1.893 | 1.00 |
+| `bedtools genomecov -bga -ibam /local/giovanni/covtobed/benchmark/../TE/example2.bam > /dev/null` | 34.897 ± 0.849 | 34.343 | 36.610 | 18.73 ± 0.48 |
