+++ benchmark/test_16.md
@@ -1,4 +1,0 @@
-| Command | Mean [ms] | Min [ms] | Max [ms] | Relative |
-|:---|---:|---:|---:|---:|
-| `covtobed ../pannelli/16.bam > /dev/null` | 964.3 ± 17.6 | 945.5 | 989.2 | 1.00 |
-| `bedtools genomecov -bga -ibam ../pannelli/16.bam > /dev/null` | 33113.9 ± 1568.7 | 30851.1 | 35317.6 | 34.34 ± 1.74 |
