+++ benchmark/test_17.md
@@ -1,4 +1,0 @@
-| Command | Mean [s] | Min [s] | Max [s] | Relative |
-|:---|---:|---:|---:|---:|
-| `covtobed ../pannelli/17.bam > /dev/null` | 1.938 ± 0.052 | 1.858 | 1.995 | 1.00 |
-| `bedtools genomecov -bga -ibam ../pannelli/17.bam > /dev/null` | 36.529 ± 0.639 | 35.866 | 37.586 | 18.85 ± 0.60 |
