### 2020-02-05 f156d23
Delete test.sh
