+++ benchmark/benchmark2_example1.md
@@ -1,0 +1,5 @@
+| Command | Mean [s] | Min [s] | Max [s] | Relative |
+|:---|---:|---:|---:|---:|
+| ` mosdepth "/local/giovanni/covtobed/benchmark"/mosd2_example1 /local/giovanni/covtobed/benchmark/../TE/example1.bam` | 67.373 ± 0.709 | 66.885 | 68.187 | 64.03 ± 0.70 |
+| `covtobed /local/giovanni/covtobed/benchmark/../TE/example1.bam > "/local/giovanni/covtobed/benchmark"/covt2_example1.bed` | 1.052 ± 0.003 | 1.050 | 1.056 | 1.00 |
+| `bedtools genomecov -bga -ibam /local/giovanni/covtobed/benchmark/../TE/example1.bam > "/local/giovanni/covtobed/benchmark"/bedt2_example1.bed` | 32.854 ± 0.177 | 32.650 | 32.961 | 31.22 ± 0.19 |
