+++ benchmark/stream/benchmarkStream_example2.md
@@ -1,0 +1,4 @@
+| Command | Mean [s] | Min [s] | Max [s] | Relative |
+|:---|---:|---:|---:|---:|
+| `covtobed /local/giovanni/covtobed/benchmark/../TE/example2.bam > /dev/null` | 1.951 ± 0.080 | 1.872 | 2.074 | 1.00 |
+| `bedtools genomecov -bga -ibam /local/giovanni/covtobed/benchmark/../TE/example2.bam > /dev/null` | 36.709 ± 0.878 | 35.784 | 38.083 | 18.82 ± 0.90 |
