+++ benchmark/get_datasets.sh
@@ -16,7 +16,7 @@ wget -O "example2.bam" "https://zenodo.org/record/3643386/files/17.bam?download=
 if  [ "NO$1" != "NO" ]; then
  echo "Get exome"
  wget "ftp://ftp.1000genomes.ebi.ac.uk/vol1/ftp/data_collections/1000_genomes_project/data/GBR/HG00258/exome_alignment/HG00258.alt_bwamem_GRCh38DH.20150826.GBR.exome.cram"
- wget "ftp://ftp.1000genomes.ebi.ac.uk/vol1/ftp/data_collections/1000_genomes_project/data/GBR/HG00258/exome_alignment/HG00258.alt_bwamem_GRCh38DH.20150826.GBR.exome.cram.cai"
+ wget "ftp://ftp.1000genomes.ebi.ac.uk/vol1/ftp/data_collections/1000_genomes_project/data/GBR/HG00258/exome_alignment/HG00258.alt_bwamem_GRCh38DH.20150826.GBR.exome.cram.crai"
  echo "Convert exome"
  samtools view HG00258.alt_bwamem_GRCh38DH.20150826.GBR.exome.cram -b > HG00258.bam
  samtools index HG00258.bam
