### 2020-02-06 c497987
Update paper.md
