+++ paper/paper.md
@@ -28,44 +28,48 @@ bibliography: paper.bib
 
 A common task in bioinformatics is the mapping of DNA sequencing reads (produced by "next generation sequencing" experiments) against a reference genome. 
 The output of the alignment is commonly encoded in a BAM file [@samformat]. 
-For several applications of DNA sequencing it is useful to extract the **depth of coverage** [@coverage] 
+For several applications of DNA sequencing it is useful to extract the *depth of coverage* [@coverage] 
 at specific positions in the BAM file, 
 encoding the output in the standard BED format [@bedtools].
 
-Here we describe `covtobed`, a C++ program designed to extract the depth of coverage per position from a sorted BAM file, 
+Here we describe *covtobed*, a C++ program designed to extract the depth of coverage per position from a sorted BAM file, 
 optionally specifying a range of coverage of interest and a minimum length for the features to be printed in the output BED file. 
 Parsing of BAM files is performed using `libbamtools` [@bamtools]. 
 
-The design has been inspired by the UNIX programming philosophy [@phylosophy], and thus `covtobed` performs a single task and supports input and output streams.
+The design has been inspired by the UNIX programming philosophy [@phylosophy], and thus *covtobed* performs a single task and supports input and output streams.
 
 # Availability and Installation
 
-`covtobed` is distributed with MIT licence and available from the [GitHub repository](https://github.com/telatin/covtobed), 
+*covtobed* is distributed with MIT licence and available from the [GitHub repository](https://github.com/telatin/covtobed), 
 and can be easily installed via Miniconda from the "bioconda" channel 
 (*i. e.* `conda install -c bioconda covtobed`).
 
 The tool is also available as a Docker image downloadable from [Docker Hub](https://hub.docker.com/r/andreatelatin/covtobed) 
-(*i. e.* `docker pull andreatelatin/covtobed`) or as a Singularity image (*i. e.* `singularity pull docker://andreatelatin/covtobed`).
+(*i. e.* `docker pull andreatelatin/covtobed`) or as a Singularity image.
 
 # Code (structure and dependencies)
 
 The code is object oriented, including an *Input* class handling reading, parsing and filtering of alignments and an *Output* class handling coverage filtering and writing in different formats.
 The main algorithm is based on a *priority_queue* from the standard library and is both fast and memory efficient.
 
-`covtobed` relies on  `libbamtools` [@bamtools] for BAM file parsing, and `cpp-optparse` [@opt] for command line option parsing.
+*covtobed* relies on  `libbamtools` [@bamtools] for BAM file parsing, and `cpp-optparse` [@opt] for command line option parsing.
 
 # Example application
 
-When performing target enrichment experiments (where the aim is to sequence a set of selected regions of a genome), it's important to detect a lack of coverage or unsufficient coverage (*i.e.* is lower than `THRESHOLD` in the target regions). This can be calculated intersecting (with `bedtools`, [@bedtools]) the target BED file with the output of `covtobed`. 
+When performing target enrichment experiments (where the aim is to sequence a set of selected regions of a genome), it's important to detect a lack of coverage or unsufficient coverage (*i.e.* is lower than `THRESHOLD` in the target regions). This can be calculated intersecting (with *bedtools*, [@bedtools]) the target BED file with the output of *covtobed*. 
 
 The tool has been used, for example, in the setup of a *target enrichment* panel targeting 71 human genes [@poloni], in order to detect uncovered regions.
 
-While a tool exists – called `mosdepth` [@mosdepth] – to perform a coverage analysis, `covtobed` was designed with the ability to quickly extract regions between used defined coverage intervals and, more importantly, with streaming from standard input and to standard output, that Mosdepth doesn't support.
[…]
