+++ paper/paper.bib
@@ -120,3 +120,15 @@
   howpublished = {\url{https://github.com/weisslj/cpp-optparse}},
   commit = {2ec0b7aca9a692ff93017ed44ca9d13a8e7d4d00}
 }
+
+@dataset{dataset,
+  author       = {Birolo, Giovanni and
+                  Telatin, Andrea},
+  title        = {{covtobed test files - 2 BAM files produced by 
+                   target enrichment experiments}},
+  month        = feb,
+  year         = 2020,
+  publisher    = {Zenodo},
+  doi          = {10.5281/zenodo.3643386},
+  url          = {https://doi.org/10.5281/zenodo.3643386}
+}
