+++ benchmark/benchmark_stream.sh
@@ -29,7 +29,8 @@ do
 	fi
 	echo "$TAG [$FILE]"
 	hyperfine --warmup 1 --min-runs 6 --cleanup 'rm *.bed || true' \
-		--export-csv benchmarkStream_$TAG.csv --export-markdown benchmarkStream_$TAG.md \
+		--export-csv stream/benchmark_$TAG.csv --export-markdown stream/benchmark_$TAG.md \
+		"samtools depth -a $FILE > /dev/null" \
 		"covtobed $FILE > /dev/null" \
 		"bedtools genomecov -bga -ibam $FILE > /dev/null"
 done
