+++ paper/paper.md
@@ -64,13 +64,13 @@ While a tool exists – called *mosdepth* [@mosdepth] – to perform a coverage 
 
 # Performance
 
-*covtobed* is orders of magnitude faster than the existing tools. In particular, *bedtools* is 20-30 times slower, and *mosdepth* is approximately 60 times slower (but will produce a compressed and indexed output). We report the measured time using a target enrichment panel [@dataset]:
+*covtobed* is faster than its alternatives. In particular, *bedtools* is approximately 20 times slower, and *mosdepth* is approximately 30 times slower (but will produce a compressed and indexed output). We report the measured time using a target enrichment panel [@dataset]:
 
 | Command | Mean [s] | Min [s] | Max [s] | Relative |
 |:---|---:|---:|---:|---:|
-| mosdepth | 67.692 ± 1.221 | 66.288 | 68.509 | 32.89 ± 0.60 |
-| covtobed | 2.058 ± 0.006 | 2.054 | 2.065 | 1.00 |
-| bedtools | 37.677 ± 0.962 | 36.568 | 38.281 | 18.31 ± 0.47 |
+| **mosdepth** | 66.626 ± 0.747 | 65.366 | 67.531 | 64.62 ± 1.52 |
+| **covtobed** | 1.031 ± 0.021 | 1.006 | 1.070 | 1.00 |
+| **bedtools** genomecov | 34.076 ± 1.644 | 33.146 | 37.360 | 33.05 ± 1.73 |
 
 The scripts to perform the benchmark are available in the [github repository](https://github.com/telatin/covtobed/tree/master/benchmark).
 
@@ -78,7 +78,6 @@ The scripts to perform the benchmark are available in the [github repository](ht
 
 The authors gratefully acknowledge the support of the Biotechnology and Biological Sciences Research Council (BBSRC); 
 this research was partly supported by the BBSRC Institute Strategic Programme Gut Microbes and Health BB/R012490/1 and its constituent project BBS/E/F/000PR10353. Analyses and benchmark performed using the MRC CLIMB cloud computing environment supported by grant MR/L015080/1.
-
 
 # References
 
