+++ paper/paper.md
@@ -64,13 +64,7 @@ While a tool exists – called *mosdepth* [@mosdepth] – to perform a coverage 
 
 # Performance
 
-*covtobed* is faster than its alternatives. In particular, *bedtools* is approximately 20-30 times slower, and *mosdepth* is approximately 30-60 times slower (but will produce a compressed and indexed output). We report the measured time using a BAM file from a target enrichment sequencing panel [@dataset]:
-
-| Command | Mean [s] | Min [s] | Max [s] | Relative |
-|:---|---:|---:|---:|---:|
-| **mosdepth** | 66.626 ± 0.747 | 65.366 | 67.531 | 64.62 ± 1.52 |
-| **covtobed** | 1.031 ± 0.021 | 1.006 | 1.070 | 1.00 |
-| **bedtools** genomecov | 34.076 ± 1.644 | 33.146 | 37.360 | 33.05 ± 1.73 |
+*covtobed* is a fast tool, outperforming the popular *bedtools* and providing comparable speed with *mosdepth* (with some datasets *covtobed* is more than ten times faster than *mosdepth*).
 
 The scripts to perform the benchmark are available in the [github repository](https://github.com/telatin/covtobed/tree/master/benchmark).
 
