### 2020-02-20 #2058 @telatin
Hello!
We tested our tool comparing its speed with *bedtools* and *mosdepth*. With either small genomes or large genomes  having a small targets (*e. g.* Human gene panels), *covtobed* outperforms both tools (up to 60X faster). In whole exome sequencing *covtobed* is 3-5X slower than *mosdepth* but 2-4X faster than *bedtools*.

We believe that this combined with peculiar features (covtobed is available for mac, allows to split the coverage by strand and works in pipes) makes the tool a useful addition in the bioinformatician's toolshed.

Hope you agree :)
