+++ README.md
@@ -10,15 +10,15 @@
 
 ### a tool to generate BED coverage tracks from BAM files
 
-Read one (or more) [alignment](https://en.wikipedia.org/wiki/Sequence_alignment) files
-(sorted [BAM](https://en.wikipedia.org/wiki/SAM_(file_format))) and prints a [BED](https://en.wikipedia.org/wiki/BED_(file_format)) with the [coverage](https://en.wikipedia.org/wiki/Coverage_(genetics)). It will join consecutive bases with the same coverage, and can be used to only print a BED file with the regions having a specific coverage range. ([Read more in the wiki]())
+Reads one (or more) [alignment](https://en.wikipedia.org/wiki/Sequence_alignment) files
+(sorted [BAM](https://en.wikipedia.org/wiki/SAM_(file_format))) and prints a [BED](https://en.wikipedia.org/wiki/BED_(file_format)) with the [coverage](https://en.wikipedia.org/wiki/Coverage_(genetics)). It will join consecutive bases with the same coverage, and can be used to only print a BED file with the regions having a specific coverage range. (:book: [Read more in the wiki]())
 
 
 ![covtobed example](img/coverage_bam_to_bed.png)
 
 ## Usage
 
-:book: The complete documentation is available in the [GitHub wiki](https://github.com/telatin/covtobed/wiki).
+:book: **The complete documentation is available in the [GitHub wiki](https://github.com/telatin/covtobed/wiki).**
 
 Synopsis:
 ```
