### 2020-02-24 #2058 @will-rowe
Hi @brentp. Would you be interested in reviewing this submission for JOSS?
