### 2020-02-25 #2058 @will-rowe
Brilliant. Thanks everyone. I'll start the review now and you'll shortly receive links to the new review issue - let me know if any of you need further instruction on the JOSS review process.
