### 2020-02-25 #2119 @whedon
[REVIEW]: covtobed: a simple and fast tool to extract coverage tracks from BAM files
**Submitting author:** @telatin (<a href="http://orcid.org/0000-0001-7619-281X">Andrea Telatin</a>)
**Repository:** <a href="https://github.com/telatin/covtobed/" target ="_blank">https://github.com/telatin/covtobed/</a>
**Version:** v1.1.0
