### 2020-02-26 telatin/covtobed#3 @brentp
JOSS tracking
re: https://github.com/openjournals/joss-reviews/issues/2119

grammar
=======
+ problems with sentence: `This can be calculated intersecting (with bedtools,`
+ "regions between used defined" should be "... user-defined" ?

benchmarking
===========
Thanks for the clear and reproducible benchmark scripts. (I have used this to reproduce your results and evaluate recent changes to mosdepth).
+ It would be more fair to also show `samtools depth` without -a as that is easily converted to a BED format.
+ either add whole-genome or a note suggesting why it's not there
+ the disk vs streaming benchmarks don't seem that informative as they have the same run-times for the softwares; is there something to make the utility more apparent?

general
======
+ samtools and mosdepth both handle CRAM files which are now a common format. It would be good to explain why this is not available in covtobed or add it as a known limitation.
+ samtools and mosdepth both handle pair overlap and avoid double-counting overlapping mates. It would be good to clearly state how this is handled in covtobed.
+ samtools and mosdepth both allow a mininum mapping quality and setting which flags are excluded/included and samtools allows a minimum base-quality. how is this handled in covtobed?
+ please add more detail about the priority queue algorithm. does it consider cigar operations or does it append the entire read?

In summary, this tool seems to nicely fill a niche--streaming coverage analysis of target gene panels (when the data is in bam format and no knobs for mapping quality or flags, etc are needed). It would be nice to expand on how covtobed is unique to that niche (relative to samtools depth and an awk one-liner to convert to bed). Additional explanation of how this works: algorithm, filtering, etc will also be helpful.
