### 2020-02-26 telatin/covtobed#3 @telatin
Thank you for the prompt feedback, it's much appreciated!
We will work to address all the questions
