### 2020-02-27 telatin/covtobed#3 @brentp
a couple more things:

I see now the mapping quality flag.

 The automated tests don't seem to test the actual output, just the line counts. Would be nice to see some that check actual coverage.

Also, I think adding more obvious notes on strand, physical coverage etc. could make the specific, unique use-cases for this tool more apparent.
