### 2020-02-27 f1ed08e
Now should work also for the end user, not only for TravisCI
