+++ benchmark/README.md
@@ -35,7 +35,6 @@ See also [example2.bam benchmark](stream/benchmarkStream_example2.md).
 
 | Command | Mean [s] | Min [s] | Max [s] | Relative |
 |:---|---:|---:|---:|---:|
-| `samtools depth -a /example1.bam` | 600.603 ± 3.205 | 596.251 | 604.440 | 573.92 ± 26.79 |
 | `bedtools genomecov -bga -ibam example1.bam` | 34.717 ± 0.952 | 33.325 | 36.090 | 33.17 ± 1.79 |
 | `covtobed example1.bam` | 1.046 ± 0.049 | 0.982 | 1.096 | 1.00 |
 
