+++ benchmark/README.md
@@ -28,16 +28,6 @@ _samtools_ is only included in the streaming section, but should be noted that p
 
 *covtobed* is up to 20 times faster than *mosdepth* on medium datasets (_e. g._ Human gene panels), while on larger datasets (_e. g._ Human whole exomes) it's up to 5 times slower.
 
-### Human panel - Straming speed (not saving to disk)
-
-_mosdepth_ is not included as it only saves to file. Other commands were redirected to `/dev/null`.
-See also [example2.bam benchmark](stream/benchmarkStream_example2.md). 
-
-| Command | Mean [s] | Min [s] | Max [s] | Relative |
-|:---|---:|---:|---:|---:|
-| `bedtools genomecov -bga -ibam example1.bam` | 34.717 ± 0.952 | 33.325 | 36.090 | 33.17 ± 1.79 |
-| `covtobed example1.bam` | 1.046 ± 0.049 | 0.982 | 1.096 | 1.00 |
-
 
 ### Human panel - Saving the output to disk
 
