### 2020-02-27 telatin/covtobed#3 @telatin
Thank you for the updates!

Before answering, I’d like to share a preamble:
 - We decided to make an effort to keep the [Wiki](https://github.com/telatin/covtobed/wiki) updated, and we would like to use it as the primary source of documentation (to avoid changing the repository to update the documentation, and to have an easy to browse document that will be probably more accessible than the paper itself, that we would like to keep concise). I expanded the Wiki with some of your suggestions
 - We used the tool in different scenarios (microbial genomics, metagenomics, human genetics), and to be of general use we decided to do a little thing in the most flexible way (input/output streaming)
To address some points:
 - *PAPER*: Thank you for the feedback on the paper: we updated it fixing the problematic bits and expanding it
 - *TESTS*: There is a test that compares the produced BED with a reference BED in the repository. Today I added a second test that will check the actual output as suggested, in this case using a very artificial BAM file making the output easy to verify.
 - *CRAM*: The limitation of not handling CRAM files has been added in the README and in the Wiki
 - *CIGAR*: covtobed doesn’t parse the CIGAR and we do not plan to do so, as our goal is to check for reagions covered by the library, not covered by actual bases (i.e. evaluate the capture efficiency). I didn’t realise, until now, that mosdepth is evaluating the CIGAR by default. We are going to clarify this, and this is why we compare the performance with “mosdepth -x”
 - *DOCUMENTATION*: I added a  second example of strand specific coverage usage in the wiki. The initial example was related to target enrichment using amplicon sequencing (to check for regions only covered by forward or by reverse reads), I added a recent usage on bacterial TraDIS.
 - *BENCHMARK*: The whole-genome benchmark has been skipped for time constraints as we don’t have examples in house, but I’m now performing the test with a publicly available dataset.
I removed samtools from the README as it can be misleading, and I added also “samtools depth” without the “-a” (only available in the subdirectory for the records).
