### 2020-02-27 9e15b7b
samtools without -a
