+++ benchmark/stream/benchmarkStream_example2.md
@@ -1,4 +1,6 @@
 | Command | Mean [s] | Min [s] | Max [s] | Relative |
 |:---|---:|---:|---:|---:|
-| `covtobed /local/giovanni/covtobed/benchmark/../TE/example2.bam > /dev/null` | 1.951 ± 0.080 | 1.872 | 2.074 | 1.00 |
-| `bedtools genomecov -bga -ibam /local/giovanni/covtobed/benchmark/../TE/example2.bam > /dev/null` | 36.709 ± 0.878 | 35.784 | 38.083 | 18.82 ± 0.90 |
+| `samtools depth -a ../TE/example2.bam > /dev/null` | 606.278 ± 1.351 | 604.367 | 607.973 | 308.35 ± 6.71 |
+| `samtools depth ../TE/example2.bam > /dev/null` | 8.779 ± 0.152 | 8.609 | 9.002 | 4.47 ± 0.12 |
+| `covtobed ../TE/example2.bam > /dev/null` | 1.966 ± 0.043 | 1.917 | 2.011 | 1.00 |
+| `bedtools genomecov -bga -ibam ../TE/example2.bam > /dev/null` | 37.296 ± 0.202 | 37.094 | 37.571 | 18.97 ± 0.42 |
