### 2020-02-27 telatin/covtobed#3 @brentp
Excellent. Thank you for explaining and updating.

There is a C++ htslib wrapper that is compatible with the bamtools API: https://github.com/38/HTSLibXX which would enable cram support. That's just a point of interest, not needed for the JOSS submission; you have addressed all my points.
