### 2020-02-29 telatin/covtobed#5 @jdeligt
Discrepancy in example output
I installed through conda and ran covtobed -m 0 -x 5 test/demo.bam on the git repo;

> EMPTY	0	414	0
NC_001416.1	0	2	0
NC_001416.1	2	6	1
NC_001416.1	6	7	2
NC_001416.1	7	12	3
NC_001416.1	12	18	4
NC_001416.1	169	170	4
NC_001416.1	201	206	4
NC_001416.1	206	207	3
NC_001416.1	207	214	4
NC_001416.1	514	539	4
NC_001416.1	539	540	3
NC_001416.1	540	569	4
NC_001416.1	596	609	4
NC_001416.1	1120	1999	0
NC_001416.1	2653	2674	4
NC_001416.1	2674	2686	3
NC_001416.1	2686	2762	2
NC_001416.1	2762	2771	1
NC_001416.1	2771	2800	0

Please update the example to match what is in the repo.
