### 2020-02-29 telatin/covtobed#6 @jdeligt
Please expand benchmarking
In my encounters with bedtools there were quite some times when output was not correct due to ambiguities of 'when to count a read' (clipped bases, Quality of read vs base, etc.). It would be good to not only benchmark performance but also concordance.

The benchmark section could benefit from a statement of what tools were considered. i.e. Artem worked on a similar re-implementation of depth analysis in sambamba, [sambamba-depth](https://github.com/biod/sambamba/wiki/%5Bsambamba-depth%5D-documentation), was this tools considered for benchmarking?
