### 2020-02-29 #2119 @jdeligt
Hi @telatin,

Overall looking good, please consider the integrity and performance check marks in;
https://github.com/telatin/covtobed/issues/6

And please update example usage as specified in;
https://github.com/telatin/covtobed/issues/5

Best regards,
Joep
