+++ README.md
@@ -57,13 +57,15 @@ covtobed -m 0 -x 5 test/demo.bam
 ```
 Output:
 ```
-NC_001416.1	1112	1120	4
-NC_001416.1	1120	1999	0
-NC_001416.1	2653	2674	4
-NC_001416.1	2674	2686	3
-NC_001416.1	2686	2762	2
-NC_001416.1	2762	2771	1
-NC_001416.1	2771	2800	0
+[...]
+NC_001416.1     0       2       0
+NC_001416.1     2       6       1
+NC_001416.1     6       7       2
+NC_001416.1     7       12      3
+NC_001416.1     12      18      4
+NC_001416.1     169     170     4
+NC_001416.1     201     206     4
+[...]
 ```
 ## Install
 
