### 2020-03-02 telatin/covtobed#5 @telatin
To keep the README short, I explicited the truncation with `[...]`.
