### 2020-03-02 53c85b1
increased coverage bits to 32
Found some bams with coverage above 65k, causing overflows in coverage!
