### 2020-03-02 telatin/covtobed#6 @telatin
Thanks for the comment, we will exapand the benchmark page.

I didn't know about sambamba-depth, do you know if it has a method to output the stats in bed format? We can add it in any case using "sambamba depth base", but the output is definitely different (per base, indeed, even if the coverage is not changing)
