+++ test/output/README.md
@@ -1,0 +1,1 @@
+Output of mosdepth, covtobed and bedtools genomecov from the demo.bam file
