+++ README.md
@@ -67,6 +67,8 @@ NC_001416.1     169     170     4
 NC_001416.1     201     206     4
 [...]
 ```
+See the full example output from different tools [:folder: here](https://github.com/telatin/covtobed/blob/master/test/output/README.md)
+
 ## Install
 
  * To install with Miniconda:
