### 2020-03-02 telatin/covtobed#6 @telatin
**Documentation**

The Wiki has been expanded with further details on the coverage calculation, explaining potential differences with other tools.
An output example has been added https://github.com/telatin/covtobed/blob/master/test/output/README.md including output from different tools.
