+++ benchmark/README.md
@@ -1,8 +1,14 @@
 # Benchmark
 
+Here we compare the speed of *covtobed* with:
+* *bedtools genomecov* - a widely used tool that will produce the same BED output as *covtobed*, exept for a different sorting of empty chromosomes
+* *mosdepth* - a powerful tool for genome coverage analysis that will also produce a BED file, although default parameters will be filtering out probably spurious alignments
+
 *covtobed* is significantly faster than *bedtools genomecov*.
 *covtobed* is faster than *mosdepth* on small genomes, and on large genomes (like the Human genome) with a limited fraction of the target covered (e. g. target enrichment panels). With panels it can be up to 60X faster than *mosdepth*.
 With large genomes highly *covered* (e. g. exomes, whole genome sequencing) is slightly slower than *mosdepth* (3-4X slower).
+
+*sambamba* also performs coverage statistics but will not print a BED file. A test has been done to compare the speed that appear slower, possibly also because of the bigger output printed. See: [comparisong with sambamba](sambamba.md)
 
 
 ![Summary](benchmark.png)
