### 2020-03-02 2a750f9
added whole genome benchmark
