### 2020-03-02 telatin/covtobed#3 @telatin
Thank again for your contributions.
I further expanded the wiki and the benchmark page with a first whole genome test and more details are about to be added.

We plan to try integrating HTSLibXX for a future update.
