### 2020-03-02 telatin/covtobed#6 @telatin
The benchmark page now has a preamble on the choice of tools, and a link to a comparison with _sambamba-depth_: https://github.com/telatin/covtobed/tree/master/benchmark#readme
