### 2020-03-02 #2119 @telatin
Dear @brentp and @jdeligt,
thank you very much for your quick and thorough review, we really feel that your input allowed us to improve the package quality.

Summarizing:
* paper updated clarifying some sentences
* automated tests have been expanded
* the wiki documentation has been expanded, specifically including information on the coverage calculation and stressing the peculiar usage of "covtobed" (physical-coverage, strand-specific analysis, streaming)
* the benchmark section has been expanded to include a test with a whole human genome and its description extended
* released 1.0.1 with support for coverage > 2**16, initially naively excluded

See: https://github.com/telatin/covtobed/issues?utf8=%E2%9C%93&q=is%3Aissue+label%3Ajoss+ for the issues opened in the repo
