+++ README.md
@@ -25,13 +25,14 @@ Reads one (or more) [alignment](https://en.wikipedia.org/wiki/Sequence_alignment
 
 Synopsis:
 ```
-Usage: covtobed [options] [BAM]...
+Usage: covtobed_mac [options] [BAM]...
 
 Computes coverage from alignments
 
 Options:
   -h, --help            show this help message and exit
   --version             show program's version number and exit
+  -v                    prints program version
   --physical-coverage   compute physical coverage (needs paired alignments in input)
   -q MINQ, --min-mapq=MINQ
                         skip alignments whose mapping quality is less than MINQ
@@ -45,7 +46,9 @@ Options:
   -l MINLEN, --min-len=MINLEN
                         print BED feature only if its length is bigger (or equal
                         to) than MINLELN (default: 1)
-  -v                    prints program version
+  -a, --only-valid-alignments
+                        skip duplicates, failed QC, and non primary alignment
+                        (default: 0)
   --output-strands      outputs coverage and stats separately for each strand
   --format=CHOICE       output format
 ```
