### 2020-03-02 telatin/covtobed#5 @jdeligt
Thanks for updating it, I was mainly referring to a line that is not just missing but discordant;
## Original
>NC_001416.1	1112	1120	4
NC_001416.1	1120	1999	0
## Actual
> NC_001416.1 596 609 4
NC_001416.1 1120 1999 0

This suggests that data has changed in the repo without the Readme being updated accordingly.
Or that functionaility changed, adding a test folder with the test output generated with the data and code in the repo can prevent things like that from happening.
I see that you have started a README in the test output. Please include the full test output and potential output files from example commaands here and populate it with a test suite or script.
