### 2020-03-02 #2119 @jdeligt
Hi @telatin ,

Thank you for the quick responses and the additions made to the repo.
I added one last comment to issue#5 but this is a nice to have but should not hold up publication.
The clarifactions will help users better determine what the software does and is a nice intro into some of the vageries of coverage depth calculations.

@will-rowe, looks like @brentp is happy with it as well so I think we'll hand it back to you.
