### 2020-03-02 telatin/covtobed#5 @telatin
I added new synthetic BAM files to allow for automated tests. In particular "test_cov.bam" has reference chromosomes named after the coverage of the only peak expected (e.g. reference _5X_ is expected to produce a single BED feature of 5X coverage).

https://github.com/telatin/covtobed/wiki/Testing
