### 2020-03-03 1e318c9
Update WGS benchmark
