### 2020-03-03 #2119 @telatin
Thanks to you all for your great work!

* **Archived release**
https://github.com/telatin/covtobed/releases/tag/v1.1.0

* **Zeonodo**
Birolo, Giovanni, & Telatin, Andrea. (2020, March 3). covtobed 1.1 - a simple and fast tool to convert the coverage track from a BAM file into a BED file (Version 1.1.0). Zenodo. http://doi.org/10.5281/zenodo.3695204
