### 2020-03-16 #2119 @arfon
> Pinging @openjournals/jose-eics - I think this can be closed now?

Just a heads up @will-rowe that you mentioned the wrong EiC group here :-)
